handle delete alert

+++ app/components/Roles/RolesCard.tsx
@@ -8,19 +8,48 @@ import { BsTrash2 } from "react-icons/bs";
 import { useDispatch, useSelector } from "react-redux";
 import { toast } from "react-toastify";
 import RolesDetailsCard from "./RolesDetailsCard";
+import Swal from "sweetalert2";
 
 const RolesCard = ({ role, isOpen, openRole, closeRole }: { role: Role; isOpen: boolean; openRole: () => void; closeRole: () => void }) => {
   const dispatch = useDispatch<AppDispatch>();
   const { singleRole } = useSelector((state: RootState) => state.roles);
   const [permissions, setPermissions] = useState<Permission[]>([]);
 
-  const handleDelete = async () => {
-    try {
-      await dispatch(deleteRole(role.id)).unwrap();
-      toast.success("Role deleted successfully");
-    } catch (err) {
-      toast.error(err as string);
-    }
+  const swalWithBootstrapButtons = Swal.mixin({
+    customClass: {
+      confirmButton: "bg-primary text-white mx-2 px-4 py-2 rounded-md",
+      cancelButton: "bg-red-500 text-white mx-2 px-4 py-2 rounded-md",
+    },
+  });
+
+  const handleDelete = () => {
+    swalWithBootstrapButtons
+      .fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonText: "Yes, delete it!",
+        cancelButtonText: "No, cancel!",
+        reverseButtons: true,
+      })
+      .then(async (result) => {
+        if (result.isConfirmed) {
+          try {
+            await dispatch(deleteRole(role.id)).unwrap();
+
+            swalWithBootstrapButtons.fire({
+              title: "Deleted!",
+              text: "Your role has been deleted.",
+              icon: "success",
+            });
+
+            toast.success("Role deleted successfully");
+          } catch (err) {
+            toast.error(err as string);
+          }
+        }
+      });
   };
 
   useEffect(() => {
